pyproject.toml
[project]
name = "llm-benchmark-script"
version = "0.3.0"
description = "deepevalを使ったローカルLLMベンチマーク比較ツール（Ollama / LM Studio 対応）"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=3.8.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.4.0",
    "openai>=1.0.0",
    "tabulate>=0.9.0",
]

[project.scripts]
llm-benchmark = "run_benchmark:main"
